+++ b2/provider.go
@@ -57,6 +57,7 @@ func New(version string, exec string) func() *schema.Provider {
 				},
 			},
 			DataSourcesMap: map[string]*schema.Resource{
+				"b2_account_info":    dataSourceB2AccountInfo(),
 				"b2_application_key": dataSourceB2ApplicationKey(),
 				"b2_bucket":          dataSourceB2Bucket(),
 				"b2_bucket_file":     dataSourceB2BucketFile(),