+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Added `b2_account_info` data source
+
 ### Changed
 * Better handling sensitive data in Terraform logs
 * Upgrade b2sdk version `>=1.4.0,<2.0.0`